+++ crates/forge/test_utils/Cargo.toml
@@ -24,6 +24,8 @@ serde.workspace = true
 tempfile.workspace = true
 tokio.workspace = true
 project-root.workspace = true
+regex.workspace = true
+snapbox.workspace = true
 forge = { path = "../../forge" }
 scarb-api = { path = "../../scarb-api" }
 forge_runner = { path = "../../forge-runner" }
