+++ crates/forge/test_utils/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod output_assert;
 pub mod runner;
 pub mod running_tests;
